requirements.txt
Flask==2.3.3
Pillow==10.0.1
piexif==1.1.3
python-dotenv==1.0.0
APScheduler==3.10.4
opencv-python==4.8.1.78
scikit-image==0.21.0
numpy==1.24.4
scikit-learn==1.3.0